move cliRestoreProgress to a separate file
